feat: added almost all functionality
Generation of DOM elements is implemented.
Pressing a key on a physical keyboard highlights the key on the virtual keyboard.
Switching keyboard layouts between.
English and another language is implemented.
Selected language saved and used on page reload.
Mouse clicks on buttons of the virtual keyboard inputs characters to the input field.
Animation of pressing a key is implemented.
Usage of ESLint